feat(fiatconnect): Handle transfer out transaction errors (#3180)
### Description

For ACT-534. Modifies our transfer out flow to direct the user to the
appropriate screen depending on the nature of the error we encounter
while attempting a transfer on the blockchain. If there is a possibility
that the transaction is in a state where it may possibly be pending on
the blockchain, we direct the user to a new "pending" screen, otherwise
we send them to the retry screen as normal. See flowchart
[here](https://www.figma.com/file/KRK3jMiv3fjrEW9XHCznKR/FC-Transfer-Out-Failure-States?node-id=0%3A1).

### Other changes

N/A

### Tested

Unit tests, and tested by manually triggering errors to fire within the
lowest-level internal send method. I have not been able to actually
produce these errors yet in a _real_ transaction on the blockchain,
however.

### How others should test

_Does this need to be tested by QA in the next release cycle? If so
please give a brief explanation of how to test these changes._

### Related issues

- Fixes ACT-534.

### Backwards compatibility

Yes.