refactor: avoid "exposing" the cloud functions base URL
We expose it specifically for Sentry, but another piece of code was using it to
construct a REST API endpoint (instead of following the convention of defining
API endpoints explicitly in networkConfig.ts). This change is pretty mundane, but
the intent it to make it harder to avoid the convention.