feat(exchanges): Add new Exchange QR code screen (#3247)
### Description

For ACT-542.

Adds in a new `ExchangeQR` screen to eventually be displayed in the CICO
Cash In flow in place of the current Exchanges screen. Uses a new QR
code generator, which, while unable to match the designs for this ticket
perfectly, does have limited capabilities for displaying a
logo/customization, which we can explore at some later date.

See [this Slack
thread](https://valora-app.slack.com/archives/C02KBT0DAHJ/p1671045146288859)
for more details on the QR code generation situation.

### Other changes

Adds in a new QR code generator; while many capabilities of this
generator are currently unused, this should set us up a bit better for
later when we want to explore our options.

### Tested

Unit and manual testing. See video below. Note that accessing the new
`ExchangeQR` screen through the button on the homescreen was only for
testing purposes; this behavior is not included in the PR.


https://user-images.githubusercontent.com/569401/207771345-930a9d01-8d44-43f3-abbf-d98478648f9c.mp4

### How others should test

N/A, as this screen is not currently used anywhere in-app.

### Related issues

- Fixes ACT-542.

### Backwards compatibility

Yes.