feat(analytics): integrate statsig experimentation with segment (#3251)
### Description

Currently, Statsig is integrated into the wallet separately from
Segment. This PR changes the way that Statsig is initialized so that the
ids used to bucket users are compatible with Segment. When events are
logged to Segment, they can now be sent to Statsig.

### Other changes

N/A

### Tested

No testing need

### How others should test

N/A

### Related issues

- ACT-467

### Backwards compatibility

N/A