chore(mtwAddress): remove accountAddress selector to prevent bugs (#3304)
### Description

Remove accountAddressSelector to prevent bugs like the one Jean fixed
[here](https://github.com/valora-inc/wallet/pull/3294/files#diff-8d33cbb0da323fba8e3986c827ba5c2d362c018756e78445f3b6c0606f413780R137).

The name of the accountAddressSelector introduces high risk of mistaking
it for a wallet address selector, but in reality, it selects MTW address
and uses wallet address as a fallback. We have largely discontinued use
of the MTW address, so this selector should not be needed anymore.

### Test plan

added unit tests

### Related issues

na

### Backwards compatibility

should be backwards compatible. assumed in 1 place that wallet address
is always non-null if MTW address is non-null-- see code comment in
`app/selectors.test.ts` for details