fix: wrongful navigation to language screen on dev reload (#3456)
### Description

Not really sure what caused this problem but it has been annoying.
During dev, after saving a file, the app will often navigate to the
language selection screen and then we'd need to force close the app to
avoid going through the onboarding flow again. This was annoying.

When kicking off the root saga, we do have a blocking saga that waits
for the redux store to be rehydrated before allowing all other sagas to
be spawn. This saga only runs if a global variable `didHydrate` is false
- between reloads, the previous value of this variable is preserved but
the store is actually rehydrated. This causes all sagas to kick off
using the un-hydrated (i.e. default) state. It is the
`checkAccountExistenceSaga` and takes us into this onboarding flow.

I added a quick fix here to always run this blocking saga in dev mode.
The rehydrate happens pretty quickly and does not introduce a meaningful
delay.

### Test plan

Tested manually

### Related issues

N/A

### Backwards compatibility

Y

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>