fix(fees): don't show error banner if fee currency balance is zero (#3488)
### Description

Estimating fees for a currency with zero balance, which results in the
"Could not calculate fees" error banner. This case accounts for ~92% of
all `estimate_fee_failed` events in the last 3 months:
https://mixpanel.com/project/2786968/view/3321744/app/insights#QXXXnHBAkKzr
The saga throws "Error: execution reverted: transfer value exceeded
balance of sender" when it tries to create a small tx and estimate gas.
Also, this banner will definitely show up for any user attempting to
cash in first time since the wallet will have a 0 balance or if the
swaps screen is loaded with 0 CELO (the default source for swaps). We
don't need to show this banner for these cases and also for any other 0
balance, since we have other checks and error messages if the user
attempts to send/swap/cashout with a currency with 0 balance.

This PR updates the feeEstimate saga to early bail if the balance is 0
and not show the banner.

### Test plan

- Unit tests
- Manually by going through the below and ensuring the banner doesn't
show up
  - CICO flow with a currency with 0 balance
  - Send flow with a currency with 0 balance
  - Swap screen with 0 CELO (the default source for swaps)


### Related issues

- Fixes part of ACT-653

### Backwards compatibility

Yes