fix(swap): inconsistent fee currency when swapping from CELO (#3491)
### Description

On the swap screen, we estimate the fee and use this to determine the
"max" "from" amount. Then when we perform the swap, we calculate the gas
price / currency again.

For a wallet with more stablecoins than CELO, when pressing "max" to
swap all CELO to something else, the "useMaxAmount" hook returns the
whole CELO balance because the stablecoin with the higher balance will
pay for gas. However when we perform the swap, CELO is chosen as the gas
currency and therefore the tx fails of course.

This PR:
- set the preferred gas currency before sending the swap, consistent
with the start of the swap flow
- increase the fee estimate buffer. i was often seeing failures when
swapping from max cEUR :/ this increased buffer seems to work better

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes RET-609

### Backwards compatibility

Y