fix: phone number format not recognised on verification screen (#3507)
### Description

Context:
https://valora-app.slack.com/archives/C029MMTD529/p1677719177909599,
https://valora-app.slack.com/archives/C025V1D6F3J/p1678194369634539?thread_ts=1677719863.631589&cid=C025V1D6F3J

We use the value in the phone number text input to do the phone number
validation, but as soon as the validation succeeds, we add the country
code to the text input. Subsequent validations are performed with a
phone number that includes the country code, which always fails. This PR
fixes this issue.

### Test plan

This phone number can now be validated:

https://user-images.githubusercontent.com/20150449/223467814-ae6beaeb-6ccd-4dfb-8ca7-77c9fcb376e9.mp4

We can also backspace on a validated number:

https://user-images.githubusercontent.com/20150449/223469228-5e1232c3-de01-49d8-9cee-4020ad71a903.mp4



### Related issues

- Fixes RET-553, ENG-29, RET-646

### Backwards compatibility

Y