feat(fiatconnect): add PIX account schema (#3525)
### Description

Currently, the PIX account schema is not supported in the wallet. This
PR adds the PIX account schema to the set of FiatConnect schemas, which
allows the account details screen to show the relevant inputs. This PR
does not yet add PIX account to the `SUPPORTED_FIAT_ACCOUNT_SCHEMAS`
set.


 

https://user-images.githubusercontent.com/20508929/224419726-8d79ffd8-3c9a-437a-85d5-9ff2265cb7a5.mp4




### Test plan

Added unit tests.

### Related issues

- ACT-656

### Backwards compatibility

Yes.

---------

Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>