feat(fiatconnect): support pix account (#3542)
### Description

In #3525, the PIX account schema was added and could be displayed in the
FiatDetailsScreen. In this PR, PIX account is added to the set of
supported fiat account schemas. Note that because the only supported PIX
account provider (Lovecrypto) requires a not-yet-supported kyc schema,
the PIX account details screen will be inaccessible to users until
ACT-646 is completed.

### Test plan

Updated unit test.

### Related issues

- ACT-656

### Backwards compatibility

Yes.