chore(statsig): log "started onboarding" timestamp to statsig (#3552)
### Description

Helps determine whether a user should be included in an experiment. The
motivating example is our Recovery Phrase in Onboarding experiment,
where we only want to show users the treatment if they see the
onboarding screen first. (For instance, we do not want to expose users
who have already completed onboarding before the experiment took place.)

The downside of this is we'll have to remember to update a targeting
gate when we start the experiment to some future time to get this to
work (even if it's just a few minutes).

Alternatives considered:
- before querying Statsig for experiment parameters (which emits an
exposure event), check some in-code logic first (like "is the user
currently going thru onboarding, or has already gone thru onboarding
while this experiment was on"). This would actually be pretty
challenging since the Statsig SDK doesn't give us a way of checking if
an experiment is running or not, and it would add what I'd describe as
"experiment business logic" to our app. I'd prefer a solution that
defers as much experiment biz logic as possible to Statsig.
- include existing users in the treatment group even if they won't see
the new onboarding screens, and use a funnel metric to exclude them in
the decision metric analysis. I'm actually not sure how we'd analyze the
control group in this case, but in any case, including users in a cohort
that we don't actually want to analyze would have downstream costs (need
more 1-off metrics, need to remember to use a funnel for all our
analysis, ...).

### Test plan
- added unit tests
- Tested manually to see if this would let us check a feature gate that
depends on it in getOnboardingStepValues. It worked. See exposure stream
on [this feature
gate](https://console.statsig.com/4plizaPmWwPL21ASV4QAO0/gates/dummy_start_timestamp_late_enough)

### Related issues

cc https://linear.app/valora/issue/ACT-643

### Backwards compatibility

na