feat(home-actions): add new swap screen used by home actions (#3623)
### Description

Part 1 of ACT-704. This PR adds a new swap screen that the "Swap" home
action will redirect to. The new screen removes the hamburger menu and
replaces it with a back button that redirects to the home screen.

For the same reasons as described in #3610, it is necessary to add a new
Swap screen rather than use existing one which is used in the Drawer
navigation.



https://user-images.githubusercontent.com/20508929/230185109-349a8f0c-8d05-41e2-9982-3e3763cabcac.mp4



### Test plan

Added unit tests

### Related issues

- Partially resolves ACT-704

### Backwards compatibility

Yes