chore(deps): upgrade segment, firebase (#3631)
### Description

upgrade `@react-native-firebase/*` deps from v11 to v17
upgrade `@segment/*` deps from v1 to v2

### Audit Advisory

This audit advisory should be able to be ignored by us because its part
of the `@expo/config-plugins` package which we definitely don't make use
of. (Expo is a platform to develop react native that we don't use).

### Test plan

I have manually tested the following: 
* identify, screen, track events all posted to segment with the data we
expect
* statsig client still initialized with same anonymousId as Segment uses
* values are still read correctly from firebase realtime DB
* feature flags are loaded correctly from firebase remote config
* notifications still received from firebase messaging
* deep links still work from firebase dynamic links

I recommend we re-test notifications and deeplinks in bugapalooza. 

### Related issues


https://linear.app/valora/issue/ACT-727/upgrade-segment-firebase-deps-in-wallet-repo

### Backwards compatibility

Yes

---------

Co-authored-by: jh2oman <jacobrwaterman@gmail.com>