chore: ensure statsig / analytics / i18n initialised before rendering app (#3672)
### Description

Satish alerted me to this `'Statsig', 'getExperimentParams: Error
getting params for experiment: dapps_filters_and_search', [Error: Call
and wait for initialize() to finish first.]` error that occurs on app
start every time. This error occurs because the root saga kicks off the
dapps saga which kicks off the saga to fetch the dapps list. We query
statsig for the experiment variables in this fetch dapps list saga. All
of this happens when the redux store is initialised, which is actually
before the AppInitGate.

We essentially need to sync the "side effects" and "render" parts of the
app.

This PR: 
- await sentry / analytics / i18n in app init saga
- ensure AppInitGate waits for all sagas to finish initialising before
starting the "render" part of the app
- ensure that analytics and i18n are not referenced before the
AppInitGate

### Test plan

Actually this error only happens on app load, there are subsequent
requests to fetch the dapps list which succeed. So no visible impact,
experiment should work as usual. The error is no longer logged, and the
app does not feel slower to start.

### Related issues

n/a

### Backwards compatibility

Y

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>