chore: upgrade to react native 0.67 (#3667)
### Description

As the title.

This was a pretty small upgrade, but one thing did change: we use
`<Modal>` components pretty extensively, these have children and a
`visible` prop. if `visible` is false, the Modal doesn't render its
children in real life but in the jest environment the children used to
be there. This is wrong - in many tests were were asserting that
elements inside `<Modal visible={false}>` components were present /
tappable etc, but actually in real life the `<Modal>` component wouldn't
have rendered any children. All the snapshot updates in this PR are for
this change, and several tests needed to be updated to be correct.

### Test plan

Did a quick manual test on iOS and Android, rely on CI

### Related issues

- Fixes RET-683

### Backwards compatibility

Y

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>
Co-authored-by: mpgaarciaa <96873750+mpgaarciaa@users.noreply.github.com>
Co-authored-by: Valora Bot <89419329+valora-bot@users.noreply.github.com>
Co-authored-by: valora-bot-crowdin <80050969+valora-bot-crowdin@users.noreply.github.com>