chore: upgrade to react native 0.69 (#3737)
### Description

The RN upgrades were light on the native side but the big change is
React 18. There are some [breaking
changes](https://blog.logrocket.com/upgrading-react-18-typescript/)
which I had to fix with a mixture of manual updates and dependency
upgrades. Of note:
- in the AppDelegate.mm file i had to move the `appStartedMillis` logic
into the new `prepareInitialProps` function
- the `use_flipper` function has been moved inside `use_react_native`,
the necessary flipper config is now passed through the input param for
`use_react_native`
- in React 18, implicit children is no longer allowed. some components
had `children` manually added to the prop defs
- annoying things happened with `Trans` and the output of `t` not being
accepted as JSX elements anymore. so i upgraded react-i18n which fixed
those issues, but the definition for the return type of `t` has changed.
previously, it was `string | undefined` but now it is `string | null`.
most of the manual changes were updating components to accept `null`
instead of `undefined`.
- `TransProps` also changed, the definition was a little hard to use so
i updated the invite modal components to take a string instead.
- the `@testing-library/react-hooks` lib was giving me typescript issues
with the `wrapper`, it actually seems that this lib has been combined
with `react-native-testing-library`. so i updated
`react-native-testing-library` in order to resolve the TS errors.
- updated the svg lib from v12 to v13 because to get their TS
definitions compatible with react 18. unfortunately v13 of the svg lib
has some issues on android with animated svg's, i spent some time trying
out the suggested fixes but couldn't get it to work. so i disabled the
animation on android for now (see QRTabBar.tsx)
- react-native-camera is deprecated, but rather than changing out the
lib right now i just patched it. ViewPropTypes is no longer an export of
react native but it was an import which caused the app to crash on the
QR screen. I just removed the import because proptypes don't affect
runtime.
- changes in unit test / snapshots: empty text components are now
rendered as `<Text />` rather than `<Text> </Text>`. some new lines were
also removed.
- the TokenBottomSheet and DAppsExplorerScreenSearch components use
debounce, and i had a weird time with the unit tests for these. the
upgraded testing lib started to behave weirdly with async tests (the
tests would time out and not move on), removing the async seemed to fix
this.

### Test plan

tested that debug builds run locally / flipper worked

### Related issues

- Fixes RET-683

### Backwards compatibility

y