chore(payment-request): Deprecate SendConfirmationLegacy's use in the payment request flow (#3778)
### Description

The only benefit of the old SendConfirmationLegacy screen was that it
allowed users to edit which currency they want to send. Instead of that,
payment requests now follow the following if/else rules to determine
which currency to send.

* If user has enough CUSD balance to fulfill, take them to
SendConfirmation screen for CUSD
* Else, try the same thing for CEUR
* Else, red error banner saying they have insufficient cUSD or cEUR to
complete the payment



https://github.com/valora-inc/wallet/assets/8432644/df35bda7-ed07-4438-b901-2c9299a1bba9



### Test plan

Added Unit tests. manually tested in mainnet.

### Related issues

[- Fixes #[issue number here]

](https://linear.app/valora/issue/ACT-764/payment-requests-sendconfirmationlegacy-=-sendconfirmation)

### Backwards compatibility

Can handle payment requests that originate from older app versions