feat(actions): Show scan icon in home actions experiment (#3802)
### Description

For ACT-809. Updates the home screen and our (4!!!) dapp screens to
display a QR scanner button when the user is bucketed into the Home
Actions experiment. On press, this button navigates to the QR scanner
screen and emits an analytics event instrumented with the originating
screen name. Additionally, on the home screen, when this QR scanner is
present, the Valora logo is no longer present. A new `showQrScanner`
parameter was added to the Statsig experiment config in order to support
this.

### Test plan

Unit and manual tested. See video and images below.

#### Home screen video

Please excuse the slowness/buginess from my simulator on the scan
screen.


https://github.com/valora-inc/wallet/assets/569401/1576feb9-b92a-4126-a071-d8751495757e

#### "Legacy" dapp explorer screenshot


![2023-05-24-163004_815x1726_scrot](https://github.com/valora-inc/wallet/assets/569401/ac4dd534-e9ae-47d6-be32-5a3a4079a79c)

#### "Search" dapp explorer screenshot


![2023-05-24-163054_814x1728_scrot](https://github.com/valora-inc/wallet/assets/569401/04c9af6e-d6db-4973-9d86-b7e7fbcb6759)

#### "Filter" dapp explorer screenshot


![2023-05-24-163129_815x1720_scrot](https://github.com/valora-inc/wallet/assets/569401/c7341d33-781d-4e49-b706-ac93acc7b6f1)

#### "Filter + Search" dapp explorer screenshot


![2023-05-24-162941_812x1722_scrot](https://github.com/valora-inc/wallet/assets/569401/c14eb76f-940e-4ae5-acbc-b59c7bf96ba4)

### Related issues

- Fixes ACT-809.

### Backwards compatibility

Yes.