fix: Use getLocalAccounts to fetch public address in jumpstart flow (#3905)
### Description

In the jumpstart flow, we get a private key, we setup a contractkit
instance and add the private key. We then get its public key and use it
to query the contract and claim rewards.

We started getting this error: `TypeError: undefined is not an object
(evaluating 'this.rpcCaller.call')`. Looks like something coming from
Forno? In any case, there's no need to call the node since the private
key is stored locally only.
Fixed it by switching from `getAccounts` to `getLocalAccounts` which is
preferable anyways because we don't make the unnecessary node call.

### Test plan

I've tested it locally. Here's a mainnet tx hash with a claim:
https://explorer.celo.org/mainnet/tx/0xedc7811ab3113fa6f3671bf63aace9ce59402ee388849edbfc9ba870b94a102c

---------

Co-authored-by: Silas Boyd-Wickizer <silas@valora.xyz>