fix(swap): Fix swap again flow (#3938)
### Description

After investigating the issue (see
[issue](https://linear.app/valora/issue/RET-772/[wallet]-[bug]-swap-again-button-doesnt-work-as-expected)),
the problem seems to be that when the app goes again to the
`SwapReviewScreen` is reusing the `SwapReviewScreen` component, and then
is using the already fetched values from the previous swap.

I'm missing the technical details in react of why this exactly happens,
I guessed it was because the swap again button goes to the
`SwapScreenWithBack`, keeping the component in the tree? I checked on
Flipper, and this was happening.
So, I changed it to use the `SwapScreen` without the back button (which
I also think is the best experience) and the bug seems to be fixed.

Another solution could be to change the `SwapReviewScreen` code to
support multiple uses. However I think we can keep it simpler if we
assume each `SwapReviewScreen` can be used once.


### Test plan

I was able to reproduce the error in the local emulator and after the
fix the error is gone.

### Related issues

- Fixes
[RET-722](https://linear.app/valora/issue/RET-772/[wallet]-[bug]-swap-again-button-doesnt-work-as-expected)

### Backwards compatibility

Yes

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>