feat(ethers): Refactor wallet management, add shared KeychainAccountManager, add Ethers wallet (#3920)
### Description

This PR contains a pretty big refactor of how we handle wallets and
accounts in the app. The relevant classes/interfaces are organized as
follows:
* The `ValoraWallet` interface is a _generic interface_ that
encapsulates wallet functionality (i.e., signing stuff) independent of
the underlying wallet library. Throughout the app, rather than expected
to use a `KeychainWallet` (which is contractkit-specific), we expect to
use a `ValoraWallet`. To clear up the nomenclature, I'm considering a
_wallet_ to mean "a collection of _accounts,_ tied to a single
blockchain", and _account_ to mean a single address/private key. This
means that `ValoraWallet` is able to perform signing on behalf of
multiple addresses. This was a side-effect of this _already_ being the
case for `KeychainWallet`, which contains multiple signers.
* `ValoraCeloWallet` and `ValoraEthersWallet` are implementations of
`ValoraWallet` using Celo libraries and Ethers, respectively.
`ValoraCeloWallet` is a _thin_ wrapper over the existing
`KeychainWallet`.
* `WalletManager` manages initializing and retrieving `ValoraWallet`
instances, and is initialized as a singleton.
* `KeychainAccountManager` is a new class (initialized as a singleton
within `WalletManager`), which manages saving, retrieving, and updating
accounts from the keychain, as well as locking and unlocking them. This
single `KeychainAccountManager` is shared between all `ValoraWallet`
implementations. As a result of this, account lock status is shared
amongst all wallets (which I'm pretty sure is desirable). Similarly,
`KeychainAccountManager` allows wallets to register _callbacks_ to run
when an account is added to the keychain; this means that adding an
account to wallet A will automatically add it to wallet B as well.

The high level _structure_ of these classes is:
* `WalletManager`
  * `ValoraCeloWallet` (implements `ValoraWallet`)
    * `KeychainWallet`
      * `KeychainSigner`s (many of these, one per account)
        * `KeychainAccountManager` (singleton, passed in via DI)
* `ValoraEthersWallet`s (implements `ValoraWallet`) (many of these, one
per chain)
    * `ethers.Wallet` (many of these, one per account)
    * `KeychainAccountManager` (singleton, passed in via DI)

`ValoraCeloWallet` is unfortunately quite nested; I didn't want to mess
around with the existing `KeychainWallet` implementation too much.
`KeychainWallet` and `KeychainSigner` are functionally not very
different than before, except a good deal of their logic has been pulled
out into `KeychainAccountManager` to be reused among other
`ValoraWallet` implementation; this should hopefully simplify
`KeychainWallet` and `KeychainSigner` a little bit.

Further, since the app now expects the interface of `ValoraWallet` in
place of `KeychainWallet`, certain callers have needed to be modified
somewhat to suit this new interface. This was largely straightforward,
_except_ for `KeychainWallet`'s `signTransaction` function, whose
parameters/return value are non-trivial to coerce from a Celo-specific
structure to something more general. As a result,
`ValoraWallet.signTransaction` retains the Celo-specific typing for the
time being.

### Test plan

Testing in progress...

### Related issues

- Fixes ACT-796, ACT-783.

### Backwards compatibility

Yes... pending testing :)