feat: fetch and store USD to local currency rate (#4162)
### Description

As the title says.

We were previously fetching and storing multiple rates, but only really
using the cUSD to local.
blockchain-api was actually giving us the rate for USD to local, since
it was assuming 1 cUSD = 1 USD.

The client behaves basically the same with regard to displaying amounts
in the local currency. but how it does it is more correct, once we'll
update blockchain-api to stop assuming 1 cUSD = 1 USD in the currency
conversion code there.

### Test plan

- Unit tests pass

### Related issues

- Fixes RET-830

### Backwards compatibility

Yes.