feat(viem): implement sendPayment (#4167)
### Description

Part 1 of implementing sendPayment with viem. Includes fetching fee
using a wrapper method to the existing chooseTxFeeDetails, creating the
appropriate `transfer` contract method based on whether the token is a
stable token or not and then simulating and then executing the method.
Adding standby tx, waiting for receipt and confirming tx will be in a
follow up PR

### Test plan

Unit tests, manually by setting the dummy flag to true and executing the
send flow. Example transaction:
https://explorer.celo.org/alfajores/tx/0x81bace7dbe2a426ff54884561da391bdb58db3ab6d63f3c26c30c72295edeef9

### Related issues

- Part of ACT-786

### Backwards compatibility

N/A